:bug: 1923 【计量器关联】关联计量器列表，给计量器设置环节时，配置环节后计量器的标签信息丢失。

+++ ems-module-power/ems-module-power-biz/src/main/java/cn/bitlinks/ems/module/power/service/standingbook/StandingbookServiceImpl.java
@@ -259,8 +259,11 @@ public class StandingbookServiceImpl implements StandingbookService {
         standingbook.setTypeId(Long.valueOf(updateReqVO.get("typeId")));
         standingbook.setId(Long.valueOf(updateReqVO.get("id")));
         // 修改标签信息 先删后增
-        standingbookLabelInfoMapper.delete(new LambdaQueryWrapper<StandingbookLabelInfoDO>().eq(StandingbookLabelInfoDO::getStandingbookId, standingbook.getId()));
-        createLabelInfoList(updateReqVO.get("labelInfo"), standingbook.getId());
+        if(StringUtils.isNotEmpty(updateReqVO.get(ATTR_LABEL_INFO))){
+            standingbookLabelInfoMapper.delete(new LambdaQueryWrapper<StandingbookLabelInfoDO>().eq(StandingbookLabelInfoDO::getStandingbookId, standingbook.getId()));
+            createLabelInfoList(updateReqVO.get(ATTR_LABEL_INFO), standingbook.getId());
+        }
+
         if (updateReqVO.get("stage") != null) {
             standingbook.setStage(Integer.valueOf(updateReqVO.get("stage")));
         }
